requirements.txt
Flask==3.1.1
Flask-APScheduler==1.13.1
Flask-Session==0.8.0
python-dotenv==1.1.0
psycopg2-binary==2.9.10
requests==2.32.3
resend==2.10.0
openai==1.85.0
Werkzeug==3.1.3